Config changes that seems to hit a "sweet spot" for workshop profitability and caravans from getting destroyed... at least within the "early" stage of campaign within the first year of game play

+++ CSharpSourceCode/Utilities/TORConfig.cs
@@ -18,8 +18,8 @@ namespace TOR_Core.Utilities
         public static int FakeBannerFrequency => _config.FakeBannerFrequency;
         public static int MaximumCustomResourceValue => _config.MaximumCustomResourceValue;
         public static int NumberOfTroopsPerFormationWithStandard => _config.NumberOfTroopsPerFormationWithStandard;
-        public static int YearsToEndEarlyCampaign => _config.YearsToEndEarlyCampaign;
-        public static int YearsToEndMidCampaign => _config.YearsToEndMidCampaign;
+        public static double YearsToEndEarlyCampaign => _config.YearsToEndEarlyCampaign;
+        public static double YearsToEndMidCampaign => _config.YearsToEndMidCampaign;
         public static int NumberOfMaximumLooterPartiesEarly => _config.NumberOfMaximumLooterPartiesEarly;
         public static int NumberOfMaximumLooterParties => _config.NumberOfMaximumLooterParties;
         public static int NumberOfMaximumLooterPartiesLate => _config.NumberOfMaximumLooterPartiesLate;
@@ -56,9 +56,9 @@ namespace TOR_Core.Utilities
             [XmlAttribute]
             public int NumberOfTroopsPerFormationWithStandard;
             [XmlAttribute]
-            public int YearsToEndEarlyCampaign;
+            public double YearsToEndEarlyCampaign;
             [XmlAttribute]
-            public int YearsToEndMidCampaign;
+            public double YearsToEndMidCampaign;
             [XmlAttribute]
             public int NumberOfMaximumLooterPartiesEarly;
             [XmlAttribute]
